fetch and display data in subjects file
